requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
ratelimit>=2.2.1
tqdm>=4.66.0